Enable setting the backend using VTK-m command line argument
The setting of the device used in VTK-m is following a convention that
is contrary to the convention in VTK-m. It is customary to set the
device in VTK-m using a command line argument. Setting it with an
environment variable is counter productive because you have to keep up
with the devices supported by VTK-m.