Add mean suffix to ensemble subsample
Add a suffix for the mean field in `SubsampleUncertaintyEnsemble`
instead of giving it the same name as the input. This matches the
behavior of `SubsampleUncertaintyIndependentGaussian`.

Keeping the mean field the same name as the input makes sense, but it
might add confusing when trying to set the field parameters for the
`ContourUncertain*` filters. It is also possible for the suffix to be
the empty string, in which case you will get the previous behavior.